Updated data available
amoseaton.csv now has a column to indicate whether or not the room is
generally available to walk into and sit into. Some rooms at RPI
generally are unlocked and available for improvement. This column
specifies this.

sections.json is the object taken from YACS' API that shows from a
classroom's perspective, which class sections are going to be in which
locations, which times, and with which professor.